authhelper: inject credentials for CSA
Inject the user credentials when running the CSA script on "external"
browsers (e.g. AJAX Spider) otherwise it's used the parameters instead
of the actual credentials.
Extract class to reduce duplication.
