+++ addOns/authhelper/CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 - Correct descriptions of the Zest script steps in the Authentication Report.
 - Fix loading/saving of Client Script Based Authentication through the GUI.
+- Inject user credentials into the script when running the Client Script Based Authentication browser integration.
 
 ## [0.25.0] - 2025-03-25
 ### Changed
